requirements.txt
aenum==3.1.15
aiohttp==3.9.3
aiosignal==1.3.1
ast2json==0.3
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
dataclasses==0.6
frozenlist==1.4.1
graphviz==0.20.1
gremlinpython==3.7.1
idna==3.6
importlib-resources==6.1.2
isodate==0.6.1
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.1
pycg==0.0.8
python-scalpel==1.0b0
PyYAML==6.0.1
redis==5.0.2
regex==2023.12.25
six==1.16.0
typed-ast==1.5.5
typeshed-client==2.5.1
typing-extensions==4.10.0
yarl==1.9.4
zipp==3.17.0
